docs(site): add Code tab to use-listbox-selection demo (core + harness)
Extract clean UseListboxSelectionExample (multiple prop, no mode checkbox)
from UseListboxSelectionDemo; rewrite UseListboxSelectionDemo as a harness
with a checkbox that drives the core and resets selection on mode switch via
key={String(multiple)}. Wire use-listbox-selection.mdx to show the core via
the Code tab and drop the redundant Example fence.
